+++ src/Grant/InstagramExchangeToken.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace League\OAuth2\Client\Grant;
+
+class InstagramExchangeToken extends AbstractGrant
+{
+    /**
+     * @inheritdoc
+     */
+    public function __toString()
+    {
+        return 'ig_exchange_token';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function getRequiredRequestParameters()
+    {
+        return [
+            'access_token',
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function getName()
+    {
+        return 'ig_exchange_token';
+    }
+}